+++ docs/com/uiza/samplebroadcast/UZBroadCastActivity.html
@@ -500,7 +500,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onInit(boolean&nbsp;success)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onInit</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onInit</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -513,7 +513,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onConnectionSuccess()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onConnectionSuccess</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onConnectionSuccess</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -526,7 +526,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onRetryConnection(long&nbsp;delay)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onRetryConnection</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onRetryConnection</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -540,7 +540,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
                                java.lang.String&nbsp;reason)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onConnectionFailed</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onConnectionFailed</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -553,7 +553,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onDisconnect()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onDisconnect</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onDisconnect</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -566,7 +566,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onAuthError()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onAuthError</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onAuthError</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -579,7 +579,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onAuthSuccess()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onAuthSuccess</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onAuthSuccess</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -592,7 +592,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;surfaceCreated()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>surfaceCreated</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>surfaceCreated</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -607,7 +607,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
                            int&nbsp;height)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>surfaceChanged</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>surfaceChanged</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -620,7 +620,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;surfaceDestroyed()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>surfaceDestroyed</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>surfaceDestroyed</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -633,7 +633,7 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onBackgroundTooLong()</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onBackgroundTooLong</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZBroadCastListener</code></dd>
+<dd><code>onBackgroundTooLong</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZBroadCastListener</code></dd>
 </dl>
 </li>
 </ul>
@@ -646,20 +646,20 @@ implements com.uiza.sdk.interfaces.UZBroadCastListener, android.view.View.OnClic
 <pre>public&nbsp;void&nbsp;onCameraChange(boolean&nbsp;isFrontCamera)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onCameraChange</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZCameraChangeListener</code></dd>
+<dd><code>onCameraChange</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZCameraChangeListener</code></dd>
 </dl>
 </li>
 </ul>
-<a name="onStatusChange-com.uiza.sdk.enums.RecordStatus-">
+<a name="onStatusChange-com.uiza.sdkbroadcast.enums.RecordStatus-">
 <!--   -->
 </a>
 <ul class="blockListLast">
 <li class="blockList">
 <h4>onStatusChange</h4>
-<pre>public&nbsp;void&nbsp;onStatusChange(com.uiza.sdk.enums.RecordStatus&nbsp;status)</pre>
+<pre>public&nbsp;void&nbsp;onStatusChange(com.uiza.sdkbroadcast.enums.RecordStatus&nbsp;status)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>onStatusChange</code>&nbsp;in interface&nbsp;<code>com.uiza.sdk.interfaces.UZRecordListener</code></dd>
+<dd><code>onStatusChange</code>&nbsp;in interface&nbsp;<code>com.uiza.sdkbroadcast.interfaces.UZRecordListener</code></dd>
 </dl>
 </li>
 </ul>
