+++ uzbroadcast/build.gradle
@@ -1,14 +1,14 @@
 apply plugin: 'com.android.library'
 apply plugin: 'com.github.dcendents.android-maven'
-group = 'com.github.uizaio.broadcast'
+group = 'com.github.uizaio'
 
 android {
     compileSdkVersion 29
     defaultConfig {
         minSdkVersion 21
         targetSdkVersion 29
-        versionCode 20071719
-        versionName "1.0.11"
+        versionCode 20071720
+        versionName "1.0.12"
     }
     dexOptions {
         jumboMode true
