--- uzbroadcast/src/main/java/com/uiza/sdk/events/EventSignal.java
@@ -1,5 +1,0 @@
-package com.uiza.sdk.events;
-
-public enum EventSignal {
-    UPDATE, STOP
-}
