--- uzbroadcast/src/main/java/com/uiza/sdk/interfaces/UZBroadCastListener.java
@@ -1,29 +1,0 @@
-package com.uiza.sdk.interfaces;
-
-public interface UZBroadCastListener {
-
-    void onInit(boolean success);
-
-    void onConnectionSuccess();
-
-    void onConnectionFailed(String reason);
-
-    void onRetryConnection(long delay);
-
-    void onDisconnect();
-
-    void onAuthError();
-
-    void onAuthSuccess();
-
-    void surfaceCreated();
-
-    void surfaceChanged(
-            int format, int width, int
-            height
-    );
-
-    void surfaceDestroyed();
-
-    void onBackgroundTooLong();
-}
