--- uzbroadcast/src/main/java/com/uiza/sdk/interfaces/UZCameraChangeListener.java
@@ -1,5 +1,0 @@
-package com.uiza.sdk.interfaces;
-
-public interface UZCameraChangeListener {
-    void onCameraChange(boolean isFrontCamera);
-}
