--- uzbroadcast/src/main/java/com/uiza/sdk/interfaces/UZCameraOpenException.java
@@ -1,8 +1,0 @@
-package com.uiza.sdk.interfaces;
-
-
-public class UZCameraOpenException extends RuntimeException {
-    public UZCameraOpenException(String message) {
-        super(message);
-    }
-}
