--- uzbroadcast/src/main/java/com/uiza/sdk/interfaces/UZRecordListener.java
@@ -1,9 +1,0 @@
-package com.uiza.sdk.interfaces;
-
-
-import com.uiza.sdk.enums.RecordStatus;
-
-public interface UZRecordListener {
-
-    void onStatusChange(RecordStatus status);
-}
