--- uzbroadcast/src/main/java/com/uiza/sdk/interfaces/UZTakePhotoCallback.java
@@ -1,7 +1,0 @@
-package com.uiza.sdk.interfaces;
-
-import android.graphics.Bitmap;
-
-public interface UZTakePhotoCallback {
-    void onTakePhoto(Bitmap bitmap);
-}
