+++ uzbroadcast/src/main/java/com/uiza/sdkbroadcast/events/EventSignal.java
@@ -1,0 +1,5 @@
+package com.uiza.sdkbroadcast.events;
+
+public enum EventSignal {
+    UPDATE, STOP
+}
