+++ uzbroadcast/src/main/java/com/uiza/sdkbroadcast/interfaces/UZCameraChangeListener.java
@@ -1,0 +1,5 @@
+package com.uiza.sdkbroadcast.interfaces;
+
+public interface UZCameraChangeListener {
+    void onCameraChange(boolean isFrontCamera);
+}
