+++ uzbroadcast/src/main/java/com/uiza/sdkbroadcast/interfaces/UZCameraOpenException.java
@@ -1,0 +1,8 @@
+package com.uiza.sdkbroadcast.interfaces;
+
+
+public class UZCameraOpenException extends RuntimeException {
+    public UZCameraOpenException(String message) {
+        super(message);
+    }
+}
