+++ uzbroadcast/src/main/java/com/uiza/sdkbroadcast/interfaces/UZRecordListener.java
@@ -1,0 +1,9 @@
+package com.uiza.sdkbroadcast.interfaces;
+
+
+import com.uiza.sdkbroadcast.enums.RecordStatus;
+
+public interface UZRecordListener {
+
+    void onStatusChange(RecordStatus status);
+}
