+++ uzbroadcast/src/main/java/com/uiza/sdkbroadcast/interfaces/UZTakePhotoCallback.java
@@ -1,0 +1,7 @@
+package com.uiza.sdkbroadcast.interfaces;
+
+import android.graphics.Bitmap;
+
+public interface UZTakePhotoCallback {
+    void onTakePhoto(Bitmap bitmap);
+}
